Perun user creation and Apache scripts.
Ansible script for Perun user and Apache are prepared.
Variables for user perun are in roles/common/vars/main.yml
Variables for Apache are in roles/common/vars/apache.yml

Configuration file for Apache server is in roles/common/files/perun.conf.
Variable for email address (and maybe something else) must be added to this file.