LDAP and LDAPc deployment
Configuration files for LDAPc were added.
Configuration of LDAP/LDAPc is completed.
For now, script for initialization of ldap (/home/perun/perun-ldapc/inicialize-ldap.sh) needs to be run manually.
Same for script start-ldapc.sh. Automatic inicialization and start of ldapc will be added in future after the necessary commit will be added to the perun system.